+++ src/main/java/org/bot/nullbot/function/BotConsumer.java
@@ -1,0 +1,8 @@
+package org.bot.nullbot.function;
+
+import com.mikuac.shiro.core.Bot;
+
+@FunctionalInterface
+public interface BotConsumer<T extends Bot, A extends Number, B> {
+    void accept(T bot, A a, B b);
+}