+++ CHANGELOG.md
@@ -786,4 +786,5 @@
   - 使用 BotCtxUtil 存储当前会话群配置等信息；
   - 新增 注解 @BotContext 用于通过 AOP 自动注册会话信息；
 - 移除 CommandProcessor 组件的 Async 注解；
+- BotFunction 函数式接口改为 BotConsumer 风格；
 - 优化部分代码。